Create Program

+++ PToP-08-Lecture/Program/ex08-02/src/App.java
@@ -1,8 +1,8 @@
 import java.util.Scanner;
 
 public class App {
-  static boolean isPrime(int N) {
-    for (int i = 2; i <= N - 1; i++) {
+  static boolean isPrime(long N) {
+    for (long i = 2; i <= N - 1; i++) {
       if (N % i == 0)
         return false;
     }
@@ -12,7 +12,7 @@ public class App {
   public static void main(String[] args) throws Exception {
     Scanner scanner = new Scanner(System.in);
     System.out.println("素数判定する正整数Nを入力してください．");
-    int N = scanner.nextInt();
+    long N = scanner.nextLong();
     scanner.close();
 
     if (isPrime(N)) {
